requirements.txt
python-dotenv
fastapi
uvicorn[standard]
python-dotenv
opencv-python
